requirements.txt
python-binance>=1.0.35
python-dotenv>=1.0.0
pandas>=3.0.0
numpy>=2.0.0
ta>=0.11.0
schedule>=1.2.2
rich>=13.7.0
requests>=2.31.0
